# docs/api/openapi-v1.yaml
openapi: 3.1.0
info:
  title: InvoiceOps API
  version: 1.0.0
  description: Contract for bilingual invoice and payment ticket triage.
servers:
  - url: http://localhost:8000
...
paths:
  /api/v1/auth/token:
    post:
      tags: [Authentication]
      summary: Issue a local PoC JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: {type: string, minLength: 1}
                password: {type: string, minLength: 1}
      responses:
        '200':
          description: Bearer token
          content:
            application/json:
              schema:
                type: object
                required: [access_token, token_type, role]
                properties:
                  access_token: {type: string}
                  token_type: {type: string, const: bearer}
                  role: {type: string, enum: [admin, reviewer, observer]}
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/classifications:
    post:
      tags: [Classification]
      summary: Classify one supplier ticket
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClassificationRequest'
            example:
              request_id: 9db8b9b1-7d2c-4d0e-8d4e-4f936d177a01
              source: email
              text: 请输入工单文本。
              taxonomy_version: invoiceops-v1
              metadata:
                channel: supplier_portal
      responses:
        '200':
          description: Classification result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClassificationResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/IdempotencyConflict'
  /api/v1/batches:
    post:
      tags: [Batches]
      summary: Submit a UTF-8 CSV batch
      security: [{bearerAuth: []}]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: UTF-8 CSV matching docs/api/batch-template.csv.
      responses:
        '202':
          description: Batch accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchAccepted'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/IdempotencyConflict'
  /api/v1/batches/{batch_id}:
    get:
      tags: [Batches]
      summary: Get batch progress and row errors
      security: [{bearerAuth: []}]
      parameters:
        - $ref: '#/components/parameters/BatchId'
      responses:
        '200':
          description: Batch status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchStatus'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/reviews:
    get:
      tags: [Reviews]
      summary: List tickets awaiting human review
      security: [{bearerAuth: []}]
      parameters:
        - in: query
          name: risk
          schema: {type: string, enum: [low, medium, high]}
        - in: query
          name: label_code
          schema: {type: string}
        - in: query
          name: waiting_min_seconds
          schema: {type: integer, minimum: 0}
        - in: query
          name: limit
          schema: {type: integer, minimum: 1, maximum: 100, default: 50}
      responses:
        '200':
          description: Review queue
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: {$ref: '#/components/schemas/ReviewItem'}
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/reviews/{ticket_id}/history:
    get:
      tags: [Reviews]
      summary: List immutable review revisions for a ticket
      security: [{bearerAuth: []}]
      parameters:
        - $ref: '#/components/parameters/TicketId'
      responses:
        '200':
          description: Review history
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: {$ref: '#/components/schemas/ReviewHistoryItem'}
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/reviews/{ticket_id}/decisions:
    post:
      tags: [Reviews]
      summary: Append a human review decision
      security: [{bearerAuth: []}]
      parameters:
        - $ref: '#/components/parameters/TicketId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewDecisionRequest'
      responses:
        '201':
          description: Decision appended
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewDecisionResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/admin/tickets:
    post:
      tags: [Admin Data]
      summary: Enter one ticket as an administrator
      security: [{bearerAuth: []}]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ClassificationRequest'}
      responses:
        '200':
          description: Classified ticket
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ClassificationResponse'}
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/IdempotencyConflict'
    get:
      tags: [Admin Data]
      summary: List administrator ticket summaries
      security: [{bearerAuth: []}]
      parameters:
        - in: query
          name: request_id
          schema: {type: string, maxLength: 128}
        - in: query
          name: risk
          schema: {type: string, enum: [low, medium, high]}
        - in: query
          name: status
          schema: {type: string, enum: [received, classified, needs_review, reviewed]}
        - in: query
          name: limit
          schema: {type: integer, minimum: 1, maximum: 100, default: 100}
      responses:
        '200':
          description: Ticket summaries
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: {type: array, items: {$ref: '#/components/schemas/AdminTicket'}}
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/admin/tickets/{ticket_id}:
    delete:
      tags: [Admin Data]
      summary: Delete one ticket as an administrator
      security: [{bearerAuth: []}]
      parameters:
        - $ref: '#/components/parameters/TicketId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Deletion result
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminDeleteResponse'}
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/IdempotencyConflict'
  /api/v1/admin/tickets/batch-delete:
    post:
      tags: [Admin Data]
      summary: Delete multiple tickets as an administrator
      security: [{bearerAuth: []}]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AdminBatchDeleteRequest'}
      responses:
        '200':
          description: Batch deletion result
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminDeleteResponse'}
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/IdempotencyConflict'
  /api/v1/admin/tickets/batch-overwrite:
    post:
      tags: [Admin Data]
      summary: Reclassify and replace multiple existing tickets
      security: [{bearerAuth: []}]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AdminBatchOverwriteRequest'}
      responses:
        '200':
          description: Batch overwrite result
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminBatchOverwriteResponse'}
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/IdempotencyConflict'
  /api/v1/admin/retention/cleanup:
    post:
      tags: [Admin Data]
      summary: Redact expired ticket text as an administrator
      security: [{bearerAuth: []}]
      responses:
        '200':
          description: Retention cleanup result
          content:
            application/json:
              schema:
                type: object
                required: [cleaned_tickets, policy, audit_proof]
                properties:
                  cleaned_tickets: {type: integer, minimum: 0}
                  policy: {type: string}
                  audit_proof: {type: string}
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/taxonomies/current:
    get:
      tags: [Contracts]
      summary: Get the active taxonomy and routing versions
      security: [{bearerAuth: []}]
      responses:
        '200':
          description: Active taxonomy contract
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaxonomyResponse'
  /api/v1/audit/{request_id}:
    get:
      tags: [Contracts]
      summary: Get the ordered audit chain for a request
      security: [{bearerAuth: []}]
      parameters:
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: Ordered audit events
          content:
            application/json:
              schema:
                type: object
                required: [request_id, events]
                properties:
                  request_id: {type: string}
                  events:
                    type: array
                    items: {$ref: '#/components/schemas/AuditEvent'}
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      description: Caller-generated key for safe retries.
      schema: {type: string, minLength: 1, maxLength: 255}
    BatchId:
      in: path
      name: batch_id
      required: true
      schema: {type: string, format: uuid}
...
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
    IdempotencyConflict:
      description: Idempotency key was reused with a different request body
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
  schemas:
...
    ClassificationRequest:
      type: object
      additionalProperties: false
      required: [request_id, source, text, taxonomy_version]
      properties:
        request_id: {type: string, minLength: 1, maxLength: 128}
        source: {type: string, enum: [email, portal, api, manual]}
        text: {type: string, minLength: 1, maxLength: 5000}
        taxonomy_version: {type: string, const: invoiceops-v1}
        metadata:
          type: object
          additionalProperties: false
          properties:
            channel: {type: string, enum: [supplier_portal, email, internal]}
            locale: {type: string, enum: [zh, en, mixed]}
...
    BatchAccepted:
      type: object
      required: [batch_id, status, trace_id]
      properties:
        batch_id: {type: string, format: uuid}
        status: {type: string, enum: [queued, processing, completed, failed]}
        trace_id: {type: string}
...